Rename create_database to database_utilities.
