+++ estimagic/dashboard/monitoring_app.py
@@ -11,7 +11,7 @@ from bokeh.models import Toggle
 
 from estimagic.dashboard.utilities import create_standard_figure
 from estimagic.dashboard.utilities import get_color_palette
-from estimagic.logging.create_database import prepare_database
+from estimagic.logging.database_utilities import prepare_database
 from estimagic.logging.read_database import read_new_iterations
 from estimagic.logging.read_database import read_scalar_field
