+++ estimagic/tests/logging/test_database_utilities.py
@@ -6,7 +6,7 @@ import pytest
 from pandas.testing import assert_frame_equal
 
 import estimagic.logging.update_database as upd_db
-from estimagic.logging.create_database import add_optimization_tables_to_database
+from estimagic.logging.database_utilities import add_optimization_tables_to_database
 from estimagic.logging.read_database import read_last_iterations
 from estimagic.logging.read_database import read_new_iterations
 from estimagic.logging.read_database import read_scalar_field
